ch(refactor): move tests out of server folder
